better variable manager
