+++ src/main.rs
@@ -1,15 +1,16 @@
-use utils::ProblemCap;
-use variable::Variables;
+use utils::{ProblemCap, Variables};
 
 mod parser;
 mod transpiler;
 mod utils;
-mod variable;
 
 fn main() {
     let mut vars = Variables::new();
     let mut trsp = transpiler::Transpiler::default();
-    let out = trsp.transpile("struct Test {}".to_string(), 0, &mut vars);
+    trsp.inject.inject = true;
+    trsp.inject.state.line = 2;
+    trsp.inject.state.column = 2;
+    let out = trsp.transpile("int x\nx".to_string(), 0, &mut vars);
     for p in trsp.problems {
         match p {
             ProblemCap::Error(problem) => {
@@ -23,5 +24,5 @@ fn main() {
             }
         }
     }
-    println!("{out}\n\n\n{:?}", vars);
+    println!("{out}\n\n\n");
 }
